+++ sample-resource-plugin/src/main/java/org/opensearch/sample/resource/actions/rest/create/CreateResourceRequest.java
@@ -25,21 +25,25 @@ import static org.opensearch.sample.utils.Constants.RESOURCE_INDEX_NAME;
 public class CreateResourceRequest extends ActionRequest implements DocRequest {
 
     private final SampleResource resource;
+    private final boolean shouldStoreUser;
 
     /**
      * Default constructor
      */
-    public CreateResourceRequest(SampleResource resource) {
+    public CreateResourceRequest(SampleResource resource, boolean shouldStoreUser) {
         this.resource = resource;
+        this.shouldStoreUser = shouldStoreUser;
     }
 
     public CreateResourceRequest(StreamInput in) throws IOException {
         this.resource = in.readNamedWriteable(SampleResource.class);
+        this.shouldStoreUser = in.readBoolean();
     }
 
     @Override
     public void writeTo(final StreamOutput out) throws IOException {
         resource.writeTo(out);
+        out.writeBoolean(shouldStoreUser);
     }
 
     @Override
@@ -49,6 +53,10 @@ public class CreateResourceRequest extends ActionRequest implements DocRequest {
 
     public SampleResource getResource() {
         return this.resource;
+    }
+
+    public boolean shouldStoreUser() {
+        return this.shouldStoreUser;
     }
 
     @Override