+++ spi/src/main/java/org/opensearch/security/spi/resources/sharing/ResourceSharing.java
@@ -131,7 +131,7 @@ public class ResourceSharing implements ToXContentFragment, NamedWriteable {
 
     @Override
     public String toString() {
-        return "ResourceSharing {" + ", resourceId='" + resourceId + '\'' + ", createdBy=" + createdBy + ", sharedWith=" + shareWith + '}';
+        return "ResourceSharing {" + "resourceId='" + resourceId + '\'' + ", createdBy=" + createdBy + ", sharedWith=" + shareWith + '}';
     }
 
     @Override